Moved tuple_type to separate file
