+++ hive/property.py
@@ -1,26 +1,8 @@
 from .hive import HiveMethodWrapper
 from .mixins import Stateful, Exportable, Bindable
+from .tuple_type import tuple_type
 from . import get_mode, get_building_hive
 from weakref import WeakSet
-
-def _check_tuple_type(value):
-    if isinstance(value, str):
-        return
-
-    assert isinstance(value, tuple), value
-    for entry in value:
-        _check_tuple_type(entry)
-
-
-def tuple_type(value):
-    if value is None:
-        return ()
-
-    if isinstance(value, str):
-        return (value,)
-    
-    _check_tuple_type(value)
-    return value
 
 
 class Property(Stateful, Bindable, Exportable):
